refactor: source, target 의 Position을 일급컬렉션으로 수정
`source` Position 과 `target` Position 을 Map 으로 가지는 일급컬렉션 적용
