+++ src/main/java/chess/domain/Command.java
@@ -1,17 +1,8 @@
 package chess.domain;
 
-import chess.domain.position.File;
-import chess.domain.position.Position;
-import chess.domain.position.Rank;
-import java.util.HashMap;
-import java.util.Map;
+import chess.domain.position.Positions;
 
 public class Command {
-
-    private static final int SOURCE_INDEX = 1;
-    private static final int TARGET_INDEX = 2;
-    private static final int FILE_INDEX = 0;
-    private static final int RANK_INDEX = 1;
     private static final String START = "start";
     private static final String END = "end";
     private static final String STATUS = "status";
@@ -42,20 +33,6 @@ public class Command {
         if (!command.startsWith(MOVE)) {
             throw new IllegalArgumentException("잘못된 커멘드 입니다.");
         }
-
-        String[] token = command.split(" ");
-
-        validatePosition(token[SOURCE_INDEX]);
-        validatePosition(token[TARGET_INDEX]);
-    }
-
-    private static void validatePosition(String token) {
-        char first = token.charAt(FILE_INDEX);
-        char second = token.charAt(RANK_INDEX);
-
-        if (!File.isFile(first) || !Rank.isRank(Character.getNumericValue(second))) {
-            throw new IllegalArgumentException("형식이 잘못되었거나 범위를 벗어났습니다.");
-        }
     }
 
     public boolean isEnd() {
@@ -74,24 +51,12 @@ public class Command {
         return command.equals(STATUS);
     }
 
-    public Map<String, Position> makePositions() {
+    public Positions makePositions() {
         checkMoveCommand();
-
-        Map<String, Position> positions = new HashMap<>();
 
         String[] token = command.split(" ");
 
-        positions.put("source", makePosition(token, SOURCE_INDEX));
-        positions.put("target", makePosition(token, TARGET_INDEX));
-
-        return positions;
-    }
-
-    private Position makePosition(String[] token, int index) {
-        File file = File.toFile(token[index].charAt(FILE_INDEX));
-        Rank rank = Rank.toRank(token[index].charAt(RANK_INDEX));
-
-        return Position.of(file, rank);
+        return Positions.from(token);
     }
 
     private void checkMoveCommand() {
