+++ src/main/java/chess/domain/state/White.java
@@ -6,7 +6,7 @@ import chess.domain.ChessBoard;
 import chess.domain.Command;
 import chess.domain.piece.Team;
 import chess.domain.position.Position;
-import java.util.Map;
+import chess.domain.position.Positions;
 
 public class White implements State {
 
@@ -31,9 +31,9 @@ public class White implements State {
     }
 
     private void checkTeam(Command command, ChessBoard chessBoard) {
-        Map<String, Position> positions = command.makePositions();
+        Positions positions = command.makePositions();
 
-        Position source = positions.get("source");
+        Position source = positions.getSource();
 
         Team team = chessBoard.findTeam(source);
